requirements.txt
Flask==3.0.0
flask-login==0.6.3
python-dotenv==1.0.0
mysql-connector-python==8.2.0
werkzeug==3.0.1
itsdangerous==2.1.2

# AI Chatbot (Gemini)
google-genai>=1.0.0

# Gmail API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
